+++ CHANGELOG.md
@@ -86,3 +86,4 @@
 - Fuel flow working
 - Removed instant fuel consumption, as its highly inaccurate using MAF calculations.
 - Start made on integrating obd activity into the main project
+- Moved instant obd data to the start drive activity